requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
python-multipart==0.0.12
openai==1.59.7
python-dotenv==1.0.1
sqlalchemy
google-generativeai
